requirements.txt
PlexAPI==4.13.2
tmdbapis==1.1.0
requests==2.28.2
opencv-python==4.7.0.68
pillow==9.4.0
pathvalidate==2.5.2
numpy==1.24.2
python-dotenv==0.21.1
psutil==5.9.4
GitPython==3.1.30